+++ spatialprofilingtoolbox/ondemand/__init__.py
@@ -1,2 +1,2 @@
 """The fast cell counts service."""
-__version__ = '0.13.0'
+__version__ = '0.23.0'