+++ spatialprofilingtoolbox/workflow/tabular_import/parsing/skimmer.py
@@ -2,11 +2,9 @@
 from importlib.resources import as_file
 from importlib.resources import files
 import re
-import json
 
 from psycopg2 import sql
-from psycopg2 import connect
-import pandas as pd
+from pandas import read_csv as pandas_read_csv
 
 from spatialprofilingtoolbox.workflow.tabular_import.parsing.cell_manifests import \
     CellManifestsParser
@@ -28,6 +26,8 @@ from spatialprofilingtoolbox.db.database_connection import create_database
 from spatialprofilingtoolbox.db.modify_constraints import DBConstraintsToggling
 from spatialprofilingtoolbox.db.modify_constraints import toggle_constraints
 from spatialprofilingtoolbox.db.schema_infuser import SchemaInfuser
+from spatialprofilingtoolbox.db.study_tokens import StudyCollectionNaming
+from spatialprofilingtoolbox.db.exchange_data_formats.study import StudyHandle
 from spatialprofilingtoolbox.standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
@@ -81,18 +81,18 @@ class DataSkimmer:
             logger.debug('%s %s', padded, table)
 
     @staticmethod
-    def _sanitize_token(token: str) -> str:
+    def sanitize_study_to_database_name(token: str) -> str:
+        handle = StudyHandle(handle=token, display_name_detail='')
+        token, _ = StudyCollectionNaming.strip_extract_token(handle)
         return re.sub(r'[ \-]', '_', token).lower()
 
     def _register_study(self, study_file: str) -> str:
-        with open(study_file, 'rt', encoding='utf-8') as file:
-            study = json.loads(file.read())
-            study_name = study['Study name']
+        study_name = StudyCollectionNaming.extract_study_from_file(study_file)
 
         if self._study_is_registered(study_name):
             raise ValueError('The study "%s" is already registered.', study_name)
 
-        database_name = self._sanitize_token(study_name)
+        database_name = self.sanitize_study_to_database_name(study_name)
         self._create_database(database_name)
         self._register_study_database_name(study_name, database_name)
         self._create_schema(study_name)
@@ -120,7 +120,7 @@ class DataSkimmer:
     def parse(self, _files) -> None:
         study_name = self._register_study(_files['study'])
         with as_file(files('adiscstudies').joinpath('fields.tsv')) as path:
-            fields = pd.read_csv(path, sep='\t', na_filter=False)
+            fields = pandas_read_csv(path, sep='\t', na_filter=False)
 
         toggle_constraints(
             self.database_config_file,